more type sync

+++ src/crypt.c
@@ -443,13 +443,13 @@ crypt_encrypt_chacha20_poly_buffer(LIBSSH2_SESSION * session,
                length
              */
             ret = chachapoly_crypt(&ctx->chachapoly_ctx, seqno, buf, buf,
-                                   ((u_int)buf_len) - 4, 4, ctx->encrypt);
+                                   buf_len - 4, 4, ctx->encrypt);
         }
         else {
             /* buf is full packet including size and auth tag but buf_len
                doesn't include size */
             ret = chachapoly_crypt(&ctx->chachapoly_ctx, seqno, buf, buf,
-                                   ((u_int)buf_len), 4, ctx->encrypt);
+                                   buf_len, 4, ctx->encrypt);
 
             /* the api expects the size field to already be removed
                from the decrypted packet so we'll help it out */
